feat: allow the use of custom search engines
The current system only allows a hardcoded selection of search engines
to be used.
With this change the user can use any search engine through the
search_engine variable.
The downside is that now people can misconfigure the variable which will
prevent search handler from working, but I think that's a fair trade
off.