fix(windows): don't use default args if user args were provided (#64)
Previously, the "start" and "explorer.exe" args would always be added to
the args list when on Windows, irrespective of whether the user overrode
the open_browser_args in the plugin config or not. This made it
impossible to configure a custom browser opening command not involving
"start" and "explorer.exe".

This commit changes the behavior to be more consistent with the
open_browser_app config, i.e. if the user provides args overrides, use
exactly what the user specifies and nothing else, which gives freedom
back to Windows users.